% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{antilogit}
\alias{antilogit}
\title{Antilogit function}
\usage{
antilogit(x)
}
\arguments{
\item{x}{Vector of unconstrained values}
}
\value{
Vector of constrained values
}
\description{
Calculate the antilogit for the object \code{x}, returning the corresponding
probability (range 0-1)
}
\keyword{antilogit}
