% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_agg.R
\name{cell_agg}
\alias{cell_agg}
\title{Aggregate compositional data within each cell}
\usage{
cell_agg(lc.df, K, pr.s, fec, pr.f, pr.eat, pr.est, pr.est.trt = NULL)
}
\arguments{
\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{K}{Vector with carrying capacity for each land cover type}

\item{pr.s}{Vector with juvenile survival probability for each land cover
type}

\item{fec}{Vector with mean per-individual fruit production for each land
cover type}

\item{pr.f}{Vector with mean probability of fruiting for each land cover type}

\item{pr.eat}{Vector with proportion of fruits eaten by birds for each land
cover type}

\item{pr.est}{Vector with probability of establishment for each land cover
type}

\item{pr.est.trt}{Tibble with grid id and modified establishment
probabilities for cells with ground cover treatments; default = NULL}
}
\value{
Named list with values aggregated within cells based on land cover
  types. Includes: \describe{ \item{lc.mx}{matrix(col=n.lc, row=ngrid) with
  LC proportions} \item{K.ag}{vector(length=ngrid) with total K}
  \item{K.lc}{matrix(col=n.lc, row=ngrid) with K per LC}
  \item{pr.s.ag}{vector(length=ngrid) with pr(surv)}
  \item{rel.dens}{matrix(col=n.lc, row=ngrid) with relative density among LC}
  \item{fec.ag}{vector(length=ngrid) with mn(fruit per adult)}
  \item{pr.f.ag}{vector(length=ngrid) with pr(fruit)}
  \item{pr.eat.ag}{vector(length=ngrid) with pr(eaten by bird)}
  \item{pr.est.ag}{vector(length=ngrid) with pr(establish)} }
}
\description{
This function reformats and calculates cell-means based on land cover
composition for relevant parameters.
}
\note{
If \code{!is.null(pr.est.trt)}, then the associated pr.est.ag values
  are substituted in the cells that received a relevant management
  treatments.
}
\keyword{aggregate,}
\keyword{initialize}
\keyword{premultiply,}
\keyword{set}
\keyword{up,}
