% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{cell_E}
\alias{cell_E}
\title{Aggregate compositional data within each cell}
\usage{
cell_E(lc.df, K, s.jv, fec, p.f, p.eat, p.est, p.est.trt = NULL,
  edges = "wall", method = "wt.mean")
}
\arguments{
\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions,
other covariates, and cell id info}

\item{K}{Vector \code{length=n.lc} with carrying capacity for each land cover
type}

\item{s.jv}{Vector \code{length=n.lc} with juvenile survival probability for
each land cover type}

\item{fec}{Vector \code{length=n.lc} with mean per-individual fruit
production for each land cover type}

\item{p.f}{Vector \code{length=n.lc} with mean probability of fruiting for
each land cover type}

\item{p.eat}{Vector \code{length=n.lc} with proportion of fruits eaten by
birds for each land cover type}

\item{p.est}{Vector \code{length=n.lc} with seedling establishment
probability for each land cover type}

\item{p.est.trt}{Tibble with grid id and modified establishment
probabilities for cells with ground cover treatments; default = NULL}

\item{edges}{Character taking the value of one of: \code{"wall", "sink",
"none"} where \code{"wall"} results in a dispersal probability of 0 for all
out-of-bound cells with no populations modeled, \code{"sink"} results in
dispersal of seeds to out-of-bound cells but no populations modeled, and
\code{"none"} results in dispersal of seeds and populations modeled}

\item{method}{Character, either: \code{"wt.mean", "lm"} where
\code{"wt.mean"} calculates each paramater as the weighted mean across land
cover types proportional to their coverage with the land cover specific
values stored in the parameter vectors, and \code{"lm"} calculates each
parameter in a regression with the slopes contained in each parameter
vector}
}
\value{
Named list with values aggregated within cells based on land cover
  types. Includes: \describe{ \item{\code{lc.mx}}{Matrix \code{(ncol=n.lc,
  nrow=ngrid)} with land cover proportions} \item{\code{K.E}}{Vector
  \code{length=ngrid} with total K} \item{\code{K.lc}}{Matrix
  \code{(ncol=n.lc, nrow=ngrid)} with K per land cover category}
  \item{\code{s.jv.E}}{Vector \code{length=ngrid} with pr(juvenile surv)}
  \item{\code{rel.dens}}{Matrix \code{(ncol=n.lc, nrow=ngrid)} with relative
  density among land cover categories} \item{\code{fec.E}}{Vector
  \code{length=ngrid} with mean fruit produced per adult)}
  \item{\code{p.f.E}}{Vector \code{length=ngrid} with fruiting probability}
  \item{\code{p.eat.E}}{Vector \code{length=ngrid} with proportion eaten by
  birds} \item{\code{p.est.E}}{Vector \code{length=ngrid} with seedling
  establishment probabilities} }
}
\description{
This function reformats and calculates expected cell-means based on land
cover composition for relevant parameters.
}
\note{
If \code{method="lm"}, then each parameter vector will be treated as a
  set of slopes for the covariates in lc.df with the number of covariates
  used in each regression is \code{length(param)-1} and the first element of
  \code{param} is the intercept.

If \code{!is.null(p.est.trt)}, then the associated p.est.E values are
  substituted in the cells that received a relevant management treatments.
}
\keyword{aggregate,}
\keyword{initialize}
\keyword{premultiply,}
\keyword{set}
\keyword{up,}
