% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{clear_forest}
\alias{clear_forest}
\title{Change forested land cover to open invasible}
\usage{
clear_forest(lc_ch.df, lc.df, sdd.pr, sdd.rate)
}
\arguments{
\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd.pr}{Array with \code{dim=c(i:disp.rows, j:disp.cols, k:2, n:ncell)} 
output from \code{\link{sdd_set_probs}}}

\item{sdd.rate}{Rate parameter for SDD exponential kernel}

\item{lc_chg.df}{Dataframe or tibble with cell grid indexes and a column for each forest type corresponding with the columns in \code{lc.df} specifying what proportion of forest should be converted to open invasible}
}
\value{
Array with dim(disp.rows, disp.cols, 2, ncell) where the third
  dimension contains grid id's for the neighborhood or probabilities to each
  target cell
}
\description{
Specified pixels have a specified proportion of specified forest type 
converted to open invasible habitat instead. This function updates the land
cover proportions within \code{lc.df} and adjusts the short distance 
dispersal kernels in \code{sdd.pr} since bird habitat preferences will alter
the dispersal to each cell with land cover changes. This is a one way,
deterministic change with no natural reversion to forest. The affected 
parameters \emphasis{must} be updated with a call to \code{\link{cell_agg}}.
}
\keyword{dispersal,}
\keyword{probabilities}
\keyword{probability,}
\keyword{sdd,}
