% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{cut_assign}
\alias{cut_assign}
\title{Assign cells to convert a proportion of forest to open invasible}
\usage{
cut_assign(nChg, ncell, lc.df, f.c)
}
\arguments{
\item{nChg}{Proportion of inbound cells to cut}

\item{ncell}{Number of inbound grid cells}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{f.c}{Vector of forest column indexes within \code{lc.df}}
}
\value{
Tibble \code{id.chg} with grid and inbound indexes of cells with land
  cover change and matrix \code{mx} with the change in each forest category
  and the total change to be added to open habitat
}
\description{
Randomly assigns a specified number of cells to have their forest habitat
cleared by a random proportion.
}
\keyword{change,}
\keyword{control,}
\keyword{cover}
\keyword{cut,}
\keyword{forest,}
\keyword{land}
\keyword{owners}
