% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{emulation_summary}
\alias{emulation_summary}
\title{Summarize BRT emulations}
\usage{
emulation_summary(resp = resp, brt.dir = "out/brt/")
}
\arguments{
\item{resp}{Which response summary to use (column name from \code{sens.out})}

\item{brt.dir}{\code{"out/brt/"} Directory with stored boosted regression
tree output}
}
\value{
List of three dataframes: relative influences, cross-validation
  deviances, and beta diversity of relative influences (i.e., for stability),
  each across different subsample sizes and tree complexities.
}
\description{
Summarize the output from global sensitivity analysis boosted regression
trees emulations. Based on function described in Prowse et al 2016. Reads BRT
output saved via emulate_sensitivity
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{sensitivity,}
