% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{expand_cnpy}
\alias{expand_cnpy}
\title{Expand to all combinations of canopy parameter ranges}
\usage{
expand_cnpy(Op = c(3, 7), Cl = c(3, 7), length_out = 2)
}
\arguments{
\item{Open}{Vector of min & max for open canopy categories (open invasible
and other)}

\item{Closed}{Vector of min & max for closed canopy categories (deciduous,
white pine, other evergreen, and mixed forests)}

\item{length.out}{Number of parameter values per land cover category}
}
\value{
List of vectors, each length 6
}
\description{
This function is similar to \code{\link[base]{expand.grid}} but inputs two
vectors with min and max parameter values (one set for open canopy and one
set for closed canopy) in addition to a length.out parameter, and returns a
list of vectors, with an element for each combination of land cover
parameters.
}
\keyword{expand.grid,}
\keyword{sensitivity}
