% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{expand_LCs}
\alias{expand_LCs}
\title{Expand to all combinations of land cover parameter ranges}
\usage{
expand_LCs(OpI = c(0.01, 0.99), Oth = c(0.01, 0.99), Dec = c(0.01, 0.99),
  WP = c(0.01, 0.99), Evg = c(0.01, 0.99), Mxd = c(0.01, 0.99),
  length_out = 2)
}
\arguments{
\item{OpI}{Vector of min & max for Open Invasible}

\item{Oth}{Vector of min & max for Other}

\item{Dec}{Vector of min & max for Deciduous Forest}

\item{WP}{Vector of min & max for White Pine Forest}

\item{Evg}{Vector of min & max for Evergreen Forest}

\item{Mxd}{Vector of min & max for Mixed Forest}

\item{length.out}{Number of parameter values per land cover category}
}
\value{
List of vectors, each length 6
}
\description{
This function is similar to \code{\link[base]{expand.grid}} but inputs six
vectors with min and max parameter values (one set for each land cover
category) in addition to a length.out parameter, and returns a list of
vectors, with an element for each combination of land cover parameters.
}
\keyword{expand.grid,}
\keyword{sensitivity}
