% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{expand_v}
\alias{expand_v}
\title{Expand all pairwise combinations of two vectors into one character vector}
\usage{
expand_v(x, y, sep = "_")
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{sep}{Like paste, specifies character(s) to separate vector values}
}
\value{
Vector (character)
}
\description{
This function is similar to \code{\link[base]{expand.grid}} but inputs two
vectors and returns a single character vector with the values from the two
vectors separated by "_" by default.
}
\keyword{expand.grid}
