% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{get_pt_id}
\alias{get_pt_id}
\title{Get the id of the cell containing specified coordinates}
\usage{
get_pt_id(lc.df, pt_coord)
}
\arguments{
\item{lc.df}{Dataframe or tibble with columns 'lat' and 'lon'}

\item{pt_coord}{Vector with longitude and latitude of a point}
}
\value{
id from lc.df of cell containing the point
}
\description{
Get the id of the cell containing specified coordinates
}
