% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_demography.R
\name{grow_lambda}
\alias{grow_lambda}
\title{Local population growth: lambda-based (a la Merow 2011)}
\usage{
grow_lambda(N.t, lambda.E, sdd.rate)
}
\arguments{
\item{N.t}{Vector of abundances with \code{length=ngrid}}

\item{lambda.E}{Vector with lambda for each cell with \code{length=ngrid}}

\item{sdd.rate}{Rate parameter for SDD exponential kernel}
}
\value{
Sparse tibble with grid id, starting population size, change in
  population size, and updated population size accounting for emigrants
}
\description{
This function calculates change in population size in each cell with
lambda-based proportional growth.
}
\keyword{growth}
\keyword{lambda,}
