% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{iterate_pop}
\alias{iterate_pop}
\title{Implement management and iterate buckthorn populations for one time step}
\usage{
iterate_pop(ngrid, ncell, N.0 = NULL, B.0 = NULL, g.p, lc.df, sdd,
  control.p = NULL, grd_cover.i = NULL, mech_chem.i = NULL,
  read_write = FALSE, path = NULL)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{N.0}{Array with initial population sizes, either returned from a
previous iteration or read from a stored .rds file in \code{path}}

\item{B.0}{Vector of initial seed bank abundances, either returned from a
previous iteration or read from a stored .rds file in \code{path}}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd}{Output with short distance dispersal neighborhoods created by
\code{\link{sdd_set_probs}}}

\item{control.p}{\code{NULL} NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{grd_cover.i}{\code{NULL} Dataframe with a row for each cell
implementing ground cover management, and columns \code{id} and \code{Trt}
detailing the cell ID and ground cover treatment (\code{"Cov", "Com",
"Lit"} for ground cover crop, compaction, or litter, respectively).}

\item{mech_chem.i}{\code{NULL} Dataframe with a row for each cell
implementing manual management of adults, and columns \code{id} and
\code{Trt} detailing the cell ID and manual treatment (\code{"M", "C",
"MC"} for mechanical, chemical, or mechanical and chemical, respectively).}

\item{read_write}{\code{FALSE} Read and write \code{N} and \code{B}}

\item{path}{\code{NULL} Directory for stored output. Overwrites files
(path/N.rds, path/B.rds) each iteration.}
}
\value{
Array N of abundances for each cell and age group, and vector B of
  seed bank abundances.
}
\description{
Run one time step of the simulation, possibly implementing management actions
}
\keyword{run,}
\keyword{simulate}
