% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{iterate_pop_econ}
\alias{iterate_pop_econ}
\title{Implement management and iterate buckthorn for one time step for parcels}
\usage{
iterate_pop_econ(parcel.df, pp.ls, N.0, B.0, g.p, lc.df, sdd, control.p,
  grd_cover.i = NULL, mech_chem.i = NULL, read_write = FALSE,
  path = NULL)
}
\arguments{
\item{parcel.df}{Dataframe with index columns ("id", "id.in", "id.pp"), and
columns detailing land cover proportions and grid proportions.}

\item{pp.ls}{List of length ncell where each element i identifies which
id.i$id.pp are within pixel i}

\item{N.0}{Array with initial population sizes, either returned from a
previous iteration or read from a stored .rds file in \code{path}}

\item{B.0}{Vector of initial seed bank abundances, either returned from a
previous iteration or read from a stored .rds file in \code{path}}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd}{Output with short distance dispersal neighborhoods created by
\code{\link{sdd_set_probs}}}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{grd_cover.i}{\code{NULL} Dataframe with a row for each cell
implementing ground cover management, and columns \code{id} and \code{Trt}
detailing the cell ID and ground cover treatment (\code{"Cov", "Com",
"Lit"} for ground cover crop, compaction, or litter, respectively).}

\item{mech_chem.i}{\code{NULL} List of dataframes (one for forest, one for
open), each with a row for each cell implementing manual management,and
columns \code{id} and \code{mort} detailing the cell ID and mortality rate
from manual management}

\item{read_write}{\code{FALSE} Read and write \code{N} and \code{B}}

\item{path}{\code{NULL} Directory for stored output. Overwrites files
(path/N.rds, path/B.rds) each iteration.}
}
\value{
Array N of abundances for each cell and age group, and vector B of
  seed bank abundances.
}
\description{
Run one time step of the simulation, implementing management actions at the
sub-pixel parcel level. This is specifically designed for integration with
the USDA-NIFA economic decision model, where management actions are taken by
individual parcels which may be smaller than the land cover map pixels. The
indexing is consequently different than for the other iteration functions.
}
\keyword{run,}
\keyword{simulate}
