% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_dispersal.R
\name{ldd_disperse}
\alias{ldd_disperse}
\title{Long distance dispersal}
\usage{
ldd_disperse(ncell, id.i, N.rcrt, n.ldd)
}
\arguments{
\item{ncell}{Number of inbound grid cells}

\item{id.i}{Tibble matching cell IDs. \code{id} indexes on the entire grid
while \code{id.inbd} indexes only inbound cells}

\item{N.rcrt}{\code{N.rcrt} output from \code{\link{new_seedlings}} with grid
id and number of new recruits in each cell}

\item{n.ldd}{Number of long distance dispersal events per time step}
}
\value{
Tibble with grid id and number of seeds
}
\description{
This function assigns n.ldd random long distance dispersal events across the
landcape. A single established seed is added to each target cell
}
\keyword{LDD,}
\keyword{dispersal}
