% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{Vector of constrained values}
}
\value{
Vector of unconstrained values
}
\description{
This function calculates the logit for the probability vector \code{x}, returning the corresponding unconstrained value
}
\keyword{antilogit}
