% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{make_grid}
\alias{make_grid}
\title{Generate landscape grid}
\usage{
make_grid(in.file, x. = "lon", y. = "lat", col.inc, out.file = NULL)
}
\arguments{
\item{in.file}{File (.csv) containing land cover composition and coordinates}

\item{col.inc}{Indexes of land cover (or other covariate) columns}

\item{out.file}{\code{NULL} File to store gridded output}

\item{lon}{\code{"lon"} Column name containing x (longitude) coordinates}

\item{lat}{\code{"lat"} Column name containing x (latitude) coordinates}
}
\value{
Dataframe with length(x)*length(y) rows and columns \code{x, y, x_y,
  inbd} and land cover or covariates
}
\description{
Create rectangular landscape grid with x and y coordinates, inbound
indicator, and land cover composition
}
\keyword{grid,}
\keyword{initialize}
\keyword{landscape,}
