% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_output.R
\name{make_plots_final_t}
\alias{make_plots_final_t}
\title{Save plots of final buckthorn distribution}
\usage{
make_plots_final_t(p.wd, g.p, N.final, txt = NULL)
}
\arguments{
\item{p.wd}{Directory to store file in}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{N.final}{Dataframe or tibble with processed output from
\code{\link{run_sim}}, including all columns from \code{lc.df} in addition
to \code{year}, \code{N.adult}, \code{N.sb}, and \code{N.less5}, filtered
to include only the final time step}

\item{txt}{Text to append to start of plot title}
}
\value{
None
}
\description{
This function saves plots of the final distribution of buckthorn. This
includes plots for presence/absence and abundance for both seeds (log
abundance) and adults.
}
\keyword{output}
\keyword{plots,}
\keyword{save,}
\keyword{store,}
