% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_output.R
\name{make_plots_gifs}
\alias{make_plots_gifs}
\title{Save gifs of buckthorn distribution through time}
\usage{
make_plots_gifs(p.wd, age.i, g.p, N.out)
}
\arguments{
\item{p.wd}{Directory to store file in}

\item{age.i}{Age or age range at first fruiting (e.g., \code{3} or
\code{3-5})}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{N.out}{Dataframe or tibble with processed output from
\code{\link{run_sim}}, including all columns from \code{lc.df} in addition
to \code{year}, \code{N.adult}, \code{N.sb}, and \code{N.less5}}
}
\value{
None
}
\description{
This function saves gifs of the distribution of buckthorn. This includes
plots for presence/absence and abundance for both seeds (log abundance) and
adults. There is also a gif distinguishing cells with a low density of
adults.
}
\keyword{gif,}
\keyword{output}
\keyword{plots,}
\keyword{save,}
\keyword{store,}
