% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_output.R
\name{make_plots_lc}
\alias{make_plots_lc}
\title{Save plots of initial land cover proportions}
\usage{
make_plots_lc(p.wd, lc.df, w = 10, h = 7)
}
\arguments{
\item{p.wd}{Directory to store file in}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{w}{Figure output width (inches)}

\item{h}{Figure output height (inches)}
}
\value{
None
}
\description{
This function saves plots of the initial coverage of each land cover category
}
\keyword{cover,}
\keyword{land}
\keyword{output}
\keyword{plots,}
\keyword{save,}
\keyword{store,}
