% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_demography.R
\name{new_seedlings}
\alias{new_seedlings}
\title{Seed germination & establishment}
\usage{
new_seedlings(ngrid, N.seed, B, p.E, g.D, g.B, s.B, dem.st = F, bank = T)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{N.seed}{\code{N.seed} output from \code{\link{ldd_disperse}} or
\code{\link{sdd_disperse}} with grid id and number of seeds in each cell}

\item{B}{Matrix with number of seeds in seed bank; dim=c(ngrid, tmax+1)}

\item{p.E}{Vector of establishment probabilities from
\code{\link{cell_E}}}

\item{g.D}{Probability of germinating in same year as produced}

\item{g.B}{Probability of germinating from seed bank}

\item{s.B}{Probability of surviving a year in the seed bank}

\item{dem.st}{\code{FALSE} denoting whether to include demographic
stochasticity}

\item{bank}{\code{TRUE} denoting whether to include a seed bank}
}
\value{
Tibble
}
\description{
Calculate the number of new seedlings in each cell
}
\keyword{run,}
\keyword{simulate}
