% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{run_sensitivity}
\alias{run_sensitivity}
\title{Run univariate sensitivity analyses}
\usage{
run_sensitivity(p, p.seq, n.sim, ngrid, ncell, g.p, control.p, lc.df,
  verbose = FALSE, makeGIFs = FALSE)
}
\arguments{
\item{p}{Parameter to perform the sensitivity analysis on}

\item{p.seq}{Sequence of values for the parameter}

\item{n.sim}{\code{2} Number of simulations for each parameter value}

\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}.
The value of parameter \code{p} will be updated within the loop for each
value in \code{p.seq}.}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{verbose}{\code{FALSE} Give updates for each year & process?}

\item{makeGIFs}{\code{FALSE} Make a gif for each parameter set?}
}
\value{
Success message
}
\description{
Running simulations over a set of varying parameters, where each parameter is
changed while holding all others constant. It runs in parallel using the
\code{doSNOW} and \code{foreach} packages.
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{sensitivity,}
