% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulation}
\usage{
run_sim(ngrid, ncell, g.p, lc.df, sdd.pr, N.init, control.p = NULL)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd.pr}{Array with sdd probabilities and neighborhoods}

\item{N.init}{Matrix or array with initial population sizes}

\item{control.p}{NULL or named list of buckthorn control treatment parameters}
}
\value{
Array of abundances for each cell and age group
}
\description{
This function runs the simulation. Currently, it runs all time steps, but for
the economic model, the structure will need to be slightly adjusted to run a
single time step. The initialization is separated from this function for that
reason.
}
\keyword{run,}
\keyword{simulate}
