% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_sim_lambda}
\alias{run_sim_lambda}
\title{Run lambda simulation}
\usage{
run_sim_lambda(ngrid, ncell, g.p, lambda, sdd.pr, N.init, control.p = NULL,
  verbose = F)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{lambda}{Vector of length \code{n.lc} with lambdas for each land cover
type}

\item{sdd.pr}{Array with sdd probabilities and neighborhoods created by
\code{\link{sdd_set_probs}}}

\item{N.init}{Matrix or array with initial population sizes created by
\code{\link{pop_init}}}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{verbose}{\code{TRUE} Give updates for each year & process?}
}
\value{
Matrix of abundances for each cell and time step
}
\description{
This function runs the simulation. Currently, it runs all time steps, but for
the economic model, the structure will need to be slightly adjusted to run a
single time step. The initialization is separated from this function for that
reason.
}
\keyword{lambda}
\keyword{run,}
\keyword{simulate,}
