% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_output.R
\name{save_abundances}
\alias{save_abundances}
\title{Save simulation output as .rds files}
\usage{
save_abundances(p.wd, ad.N, sb.N, verbose = TRUE)
}
\arguments{
\item{p.wd}{Directory to store file in}

\item{ad.N}{Array of adult abundances with \code{dim=c(ngrid, tmax+1, n.sim)}
generated with a call or calls to \code{\link{run_sim}}}

\item{verbose}{\code{TRUE} Confirm file creation?}

\item{ad.N}{Array of seed bank abundances with \code{dim=c(ngrid, tmax+1,
n.sim)} generated with a call or calls to \code{\link{run_sim}}}
}
\value{
None
}
\description{
This function saves both the global and buckthorn control parameters used in
a particular simulation run as a \code{.rds} file in the same directory as
the associated plots.
}
\keyword{abundances,}
\keyword{output}
\keyword{save,}
\keyword{store,}
