% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_output.R
\name{save_pars}
\alias{save_pars}
\title{Save global and control parameters as RData files}
\usage{
save_pars(p.wd, age.i, g.p, control.p)
}
\arguments{
\item{p.wd}{Directory to store file in}

\item{age.i}{Age or age range at first fruiting (e.g., \code{3} or
\code{3-5})}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}
}
\value{
None
}
\description{
This function saves both the global and buckthorn control parameters used in
a particular simulation run as a \code{.RData} file in the same directory as
the associated plots.
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{store,}
