% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdd_disperse.R
\name{sdd_disperse}
\alias{sdd_disperse}
\title{Short distance dispersal}
\usage{
sdd_disperse(id_i, N.f, pr.eat.ag, pr.s.bird, sdd.pr, sdd.rate, sdd.st = F)
}
\arguments{
\item{id_i}{Tibble matching cell IDs. \code{id} indexes on the entire grid
while \code{id_inbd} indexes only inbound cells}

\item{N.f}{Tibble of fruits produced in each cell output from
\code{\link{make_fruits}}}

\item{pr.eat.ag}{Vector of proportion of fruits eaten by birds from
\code{\link{cell_agg}}}

\item{pr.s.bird}{Proportion of viable seeds post-digestion}

\item{sdd.pr}{Array with dim(i:disp.rows, j:disp.cols, k:2, n:ncell) output
from \code{\link{sdd_set_probs}}}

\item{sdd.rate}{Rate parameter for SDD exponential kernel}

\item{sdd.st}{\code{Logical} denoting whether to implement short distance
dispersal stochastically}
}
\value{
Tibble with grid id and number of seeds in each cell
}
\description{
This function calculates the number of viable seeds deposited within the
short distance dispersal neighborhood of each cell, accounting for the
distance from the source cell, bird habitat preference, the proportion of
fruits eaten by birds, and seed viability post-digestion.
}
\keyword{SDD}
\keyword{dispersal,}
