% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_dispersal.R
\name{sdd_lambda}
\alias{sdd_lambda}
\title{Short distance dispersal: lambda-based}
\usage{
sdd_lambda(N.new, id.i, sdd.pr, sdd.rate, K.ag, sdd.st = F)
}
\arguments{
\item{id.i}{Tibble matching cell IDs. \code{id} indexes on the entire grid
while \code{id.in} indexes only inbound cells}

\item{sdd.pr}{Array with dim(i:disp.rows, j:disp.cols, k:2, n:ncell) output
from \code{\link{sdd_set_probs}}}

\item{sdd.rate}{Rate parameter for SDD exponential kernel}

\item{K.ag}{Carrying capacity of each cell output from \code{\link{cell_E}}}

\item{sdd.st}{\code{Logical} denoting whether to implement short distance
dispersal stochastically}

\item{ncell}{Number of inbound grid cells}
}
\value{
Tibble with grid id and number of individuals limited by K
}
\description{
This function calculates the number of immigrants to each cell, accounting
for distance from source cell and bird habitat preferences. It uses the
output from simple lambda-based population growth rather than the demographic
version. The output values account for carrying capacity.
}
\keyword{SDD,}
\keyword{dispersal,}
\keyword{lambda}
