% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_dispersal.R
\name{sdd_set_probs}
\alias{sdd_set_probs}
\title{Set short distance dispersal probabilities}
\usage{
sdd_set_probs(ncell, lc.df, g.p, lc.new = NULL, edges = "wall")
}
\arguments{
\item{ncell}{Number of inbound grid cells}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{g.p}{Named list of global parameters}

\item{lc.new}{Vector of indexes of cells that need to have their SDD
neighborhood recalculated; defaults to \code{NULL} and calculates SDD
neighborhoods for all cells}
}
\value{
Array with dim(disp.rows, disp.cols, 2, ncell) where the third
  dimension contains grid id's for the neighborhood or probabilities to each
  target cell
}
\description{
Assign base dispersal probabilities from each cell. Each layer \code{k} in
\code{[1:i, 1:j, k, source.id]} is the SDD neighborhood for cell n. k=1
contains pr(SDD | source.id,i,j); k=2 contains the grid id for each cell in
the neighborhood.
}
\keyword{dispersal,}
\keyword{probabilities}
\keyword{probability,}
\keyword{sdd,}
