% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_dispersal.R
\name{sdd_update_probs}
\alias{sdd_update_probs}
\title{Update short distance dispersal probabilities}
\usage{
sdd_update_probs(lc.df, g.p, sdd.alter, sdd.0, lc.col = 4:9)
}
\arguments{
\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{g.p}{Named list of global parameters}

\item{sdd.alter}{Vector of indexes of cells that need to have their SDD
neighborhood recalculated}

\item{sdd.0}{Original (non-sparse) SDD neighborhoods generated by
\link{sdd_set_probs}; i.e., sdd_set_probs()$i}

\item{lc.col}{\code{4:9} Column indexes for land cover proportions}

\item{edges}{\code{"wall"} Boundary behavior}
}
\value{
List with full neighborhoods,i, sparse representation, sp, and sparse
  dataframe sp.df. The full array has dim(disp.rows, disp.cols, 2, ncell)
  where the third dimension contains grid id's for the neighborhood or
  probabilities to each target cell. The sparse representation contains a
  list with containing the cells dispersing into each cell and a list with
  the associated probabilities. The sparse dataframe is a dataframe with a
  row for each non-zero i-j pair with columns for i, j, and dispersal
  probability
}
\description{
Update dispersal probabilities for each affected cell following land cover
changes. Because SDD probabilities are weighted by land cover based on bird
habitat preferences, the probabilities must be recalculated for each cell
that contains an affected cell as its target. Each layer \code{k} in
\code{[1:i, 1:j, k, source.id]} is the SDD neighborhood for cell n. k=1
contains pr(SDD | source.id,i,j); k=2 contains the grid id for each cell in
the neighborhood.
}
