% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{set_control_p}
\alias{set_control_p}
\title{Set buckthorn control treatment parameters}
\usage{
set_control_p(null_ctrl = TRUE, t.trt = 30, add.owners = FALSE,
  nTrt.grd = 0.05, nTrt.man = 0.05, grd.trt = c(Lit = 0.005, Cov = 0.01,
  Com = 1e-05), man.trt = c(M = 0.1, C = 0.3, MC = 0.8), lc.chg = TRUE,
  n.chg = 1e-04)
}
\arguments{
\item{null_ctrl}{\code{Logical} Set control parameters to \code{NULL}?}

\item{t.trt}{Year to start treatments}

\item{add.owners}{\code{Logical} Do owners treat every year once starting a
particular treatment?}

\item{nTrt.grd}{Proportion of cells with ground treatments in each time step}

\item{nTrt.man}{Proportion of cells with manual treatments in each time step}

\item{grd.trt}{Named vector with ground treatments and associated seedling
establishment probabilities}

\item{man.trt}{Named vector with manual treatments and associated mortality
(=success) rates}

\item{lc.chg}{\code{Logical} Does land cover change across years?}

\item{n.chg}{Proportion of cells with land cover change each year}
}
\value{
Named list of control parameters including all arguments as elements
  unless \code{null_ctrl==TRUE}, in which case the function returns
  \code{NULL}
}
\description{
This function sets the buckthorn control treatment parameters for the
simulation, allowing for individual elements to be reassigned.
}
\keyword{control,}
\keyword{initialize,}
\keyword{parameter}
\keyword{set}
\keyword{treatment,}
\keyword{up,}
