% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{set_sensitivity_pars}
\alias{set_sensitivity_pars}
\title{Set parameters for global sensitivity analysis}
\usage{
set_sensitivity_pars(pars)
}
\arguments{
\item{pars}{Names of parameters to perform sensitivity analysis on}
}
\value{
Named list with a sublist for each parameter, including the parameter
  name, the type ('prob', 'cont', 'int'), whether it varies by land cover
  category, and the minimum and maximum values
}
\description{
Generate parameter ranges and details for running a global sensitivity
analysis.
}
\keyword{initialize}
\keyword{parameters,}
\keyword{sensitivity,}
