% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_assign.R
\name{trt_assign}
\alias{trt_assign}
\title{Assign landowner treatments}
\usage{
trt_assign(id_i, ncell, nTrt, trt.eff, addOwners = FALSE, trt.m1 = NULL)
}
\arguments{
\item{id_i}{Tibble matching cell IDs. \code{id} indexes on the entire grid
while \code{id_inbd} indexes only inbound cells}

\item{ncell}{Number of inbound grid cells}

\item{nTrt}{Proportion of inbound grid cells to be treated. If add != NULL,
then this is appended to the cells treated in the previous time step.}

\item{trt.eff}{Named vector with effects of each treatment. These names are
sampled and assigned to cells}

\item{addOwners}{\code{Logical} denoting whether to append new treated cells
to the previous time step. If \code{TRUE}, then \code{trt.m1} must be
provided}

\item{trt.m1}{Tibble with grid id and treatment type from time \code{t - 1},
either empty or generated with \code{\link{trt_assign}} in the previous
time step.}
}
\value{
Tibble with grid id and treatment type
}
\description{
This function randomly assigns pixels to the specified buckthorn management
treatment. Assigned cells either overwrite the cells treated in the previous
time step or are appended.
}
\keyword{control,}
\keyword{owners}
\keyword{treatment,}
