% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{trt_econ}
\alias{trt_econ}
\title{Implement cutting & spraying treatments for econ simulations}
\usage{
trt_econ(N.t, m.max, id.trt)
}
\arguments{
\item{N.t}{Matrix or array of abundances, with dims=c(ngrid, (lc), m.max)}

\item{m.max}{Max(age at maturity)}

\item{id.trt}{Dataframe with the grid id and mortality rate for each cell}
}
\value{
Matrix or array of the same dimensions as N.t with adjusted abundances
}
\description{
Adjust the abundances within cells receiving cutting and/or spraying 
treatments according to the mortality rates calculated by the economic 
decision model.
}
\keyword{control,}
\keyword{cutting,}
\keyword{manual,}
\keyword{spraying}
\keyword{treatment,}
