% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{trt_ground}
\alias{trt_ground}
\title{Implement ground cover treatments}
\usage{
trt_ground(est.trt, grd.trt)
}
\arguments{
\item{est.trt}{Tibble output from \code{\link{trt_assign}} with the grid id
and treatment type for each cell}

\item{grd.trt}{Named vector with treatment types and associated establishment
probability}
}
\value{
Tibble with grid id and establishment probabilities
}
\description{
This function overwrites the establishment probabilities for cells receiving
a ground cover treatment
}
\keyword{compaction,}
\keyword{control,}
\keyword{cover}
\keyword{crop}
\keyword{litter,}
\keyword{manual,}
\keyword{treatment,}
