% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{trt_manual}
\alias{trt_manual}
\title{Implement cutting & spraying treatments}
\usage{
trt_manual(N.t, m.max, N.trt, man.trt)
}
\arguments{
\item{N.t}{Matrix or array of abundances, with dims=c(ngrid, (lc), m.max)}

\item{m.max}{Max(age at maturity)}

\item{N.trt}{Tibble output from \code{\link{trt_assign}} with the grid id and
treatment type for each cell}

\item{man.trt}{Named vector with treatment types and associated success
(=mortality) rates}
}
\value{
Matrix or array of the same dimensions as N.t with adjusted
  abundances
}
\description{
Adjust the abundances within cells receiving cutting and/or spraying
treatments.
}
\keyword{control,}
\keyword{cutting,}
\keyword{manual,}
\keyword{spraying}
\keyword{treatment,}
