% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_management.R
\name{cut_forest}
\alias{cut_forest}
\title{Change forested land cover to open invasible}
\usage{
cut_forest(id.chg, forest.chg, forest.col, lc.df)
}
\arguments{
\item{id.chg}{Dataframe or tibble with cell grid indexes identifying which
cells are to change. if not input manually, can be randomly created by
\code{\link{cut_assign}}}

\item{forest.chg}{Matrix with a column for each forest.col type corresponding
with the forest.col columns in \code{lc.df} specifying the amount of each
category of forest.col to be converted, and a final column \code{TotChg}
with the total amount converted to open invasible. If not input manually,
can be randomly created by \code{\link{cut_assign}}}

\item{forest.col}{Vector of forest.col column indexes within \code{lc.df}}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}
}
\value{
List with \code{lc.df} and \code{sdd.pr} where both are sparse,
  containing only updated values for the cells with altered land cover.
}
\description{
Convert forest to open habitat. The specified pixels have a specified
proportion of specified forest.col type converted to open invasible habitat.
This function updates the land cover proportions within \code{lc.df}. This is
a one way, deterministic change with no natural reversion to forest.col. The
affected parameters \emph{must} be updated with a call to
\code{\link{cell_agg}} and to \code{\link{sdd_set_probs}(lc.new=id.chg)}.
}
\keyword{change,}
\keyword{control,}
\keyword{cover}
\keyword{cut,}
\keyword{forest.col,}
\keyword{land}
\keyword{owners}
