% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{emulate_sensitivity}
\alias{emulate_sensitivity}
\title{Emulate global sensitivity analysis output}
\usage{
emulate_sensitivity(sens.out, pars, pars.rng, prop.sub, tree.depth, response,
  verbose = T)
}
\arguments{
\item{sens.out}{Dataframe of the parameter sets and simulation summaries;
\code{results} from \link{global_sensitivity}}

\item{pars}{Parameter the global sensitivity analysis included}

\item{pars.rng}{Dataframe with col(p=parameters, min=minimum, max=maximum)
listing the ranges for each parameter}

\item{prop.sub}{Proportion of sensitivity analysis output to use}

\item{tree.depth}{Vector of regression tree interaction depths to test}

\item{response}{Which response summary to use (column name from
\code{sens.out})}

\item{verbose}{\code{FALSE} Give updates for each year & process?}
}
\value{
Success message
}
\description{
Emulate the output from a global sensitivity analysis using boosted
regression trees with different interaction depths. Based on function
described in Prowse et al 2016.
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{sensitivity,}
