% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{emulate_sensitivity}
\alias{emulate_sensitivity}
\title{Emulate global sensitivity analysis output}
\usage{
emulate_sensitivity(sens.out, par.ls, n.cores = 1, n.sub = 10,
  td = c(1, 3, 5), resp, brt.dir = "out/brt/")
}
\arguments{
\item{sens.out}{Dataframe of the parameter sets and simulation summaries;
\code{.$results} from \link{global_sensitivity}}

\item{par.ls}{List output from \code{\link{set_par_ranges}} with parameter to
perform the sensitivity analysis on}

\item{n.cores}{\code{1} Number of cores for fitting subsample BRTs}

\item{n.sub}{\code{10} Number of subsamples for each emulation}

\item{td}{\code{c(1,3,5)} Vector of regression tree interaction depths to
test}

\item{resp}{Which response summary to use (column name from \code{sens.out})}

\item{brt.dir}{\code{"out/brt/"} Directory to store boosted regression tree
output}
}
\value{
Success message
}
\description{
Emulate the output from a global sensitivity analysis using boosted
regression trees with different interaction depths. Based on function
described in Prowse et al 2016. Writes files to out/brt, creating /brt if
necessary
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{sensitivity,}
