% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sensitivity.R
\name{global_sensitivity}
\alias{global_sensitivity}
\title{Run global sensitivity analysis}
\usage{
global_sensitivity(par.ls, nSamp, ngrid, ncell, g.p, lc.df, sdd = NULL,
  N.init, control.p = NULL, verbose = FALSE, sim.dir = "out/sims/")
}
\arguments{
\item{par.ls}{List output from \code{\link{set_par_ranges}} with parameter to
perform the sensitivity analysis on}

\item{nSamp}{\code{10} Number of randomly generated parameter sets}

\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}.
The value of parameter \code{p} will be updated within the loop for each
value in \code{p.seq}.}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd}{\code{NULL} Output with short distance dispersal neighborhoods
created by \code{\link{sdd_set_probs}}. If \code{NULL}, sdd neighborhoods
are calculated}

\item{N.init}{Matrix or array with initial population sizes created by
\code{\link{pop_init}}}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{verbose}{\code{FALSE} Give updates?}

\item{sim.dir}{\code{"out/sims/"} Directory to store simulation output}
}
\value{
list with elements \code{out} containing the full output from
  \link{run_sim}, and \code{results} containing the parameter sets and
  simulation summaries
}
\description{
Generate and run simulations over a set of varying parameters. It runs in
parallel using the \code{snow} and \code{foreach} packages.
}
\keyword{output}
\keyword{parameters,}
\keyword{save,}
\keyword{sensitivity,}
