% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_demography.R
\name{make_fruits}
\alias{make_fruits}
\title{Local fruit production}
\usage{
make_fruits(N.t, lc.mx, fec.E, pr.f.E, y.ad, age.f.d, dem.st = F)
}
\arguments{
\item{N.t}{Matrix or array of abundances, with dims=c(ngrid, (lc), y.ad)}

\item{lc.mx}{Matrix of land cover proportions from \code{\link{cell_E}}}

\item{fec.E}{Vector of fruit per individual from \code{\link{cell_E}}}

\item{pr.f.E}{Vector of fruiting probability from \code{\link{cell_E}}}

\item{y.ad}{Maximum age at maturity (i.e., \code{max(age.f)})}

\item{age.f.d}{\code{Logical} denoting whether \code{age.f} differs among
land cover types}

\item{dem.st}{\code{Logical} denoting whether to include demographic
stochasticity}
}
\value{
Tibble with grid id, number of reproducing individuals, and total
  fruit produces within the cell
}
\description{
This function calculates the number of fruits produces in ecah grid cell. It
assumes no fruit production before \code{age.f} and that individuals are
distributed among land cover types in densities relative to K.
}
\keyword{fecundity}
\keyword{fruit,}
\keyword{reproduction,}
