% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_demography.R
\name{new_seedlings}
\alias{new_seedlings}
\title{Seed germination & establishment}
\usage{
new_seedlings(ngrid, N.seed, N.sb, pr.est.ag, pr.sb, dem.st = F, bank = F)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{N.seed}{\code{N.seed} output from \code{\link{ldd_disperse}} or
\code{\link{sdd_disperse}} with grid id and number of seeds in each cell}

\item{N.sb}{Matrix with number of seeds in seed bank; dim=c(ngrid, tmax+1)}

\item{pr.est.ag}{Vector of establishment probabilities from
\code{\link{cell_agg}}}

\item{pr.sb}{Probability of surviving a year in the seed bank}

\item{dem.st}{\code{Logical} denoting whether to include demographic
stochasticity}

\item{bank}{\code{Logical} denoting whether to include a seed bank}
}
\value{
Tibble
}
\description{
This function calculates the number of new seedlings in each cell
}
\keyword{run,}
\keyword{simulate}
