% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_sim}
\alias{run_sim}
\title{Run demographic simulation}
\usage{
run_sim(ngrid, ncell, g.p, lc.df, sdd, N.init, control.p, verbose = TRUE,
  save_yrs = NULL, K_max = NULL)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd}{Output with short distance dispersal neighborhoods created by
\code{\link{sdd_set_probs}}}

\item{N.init}{Matrix or array with initial population sizes created by
\code{\link{pop_init}}}

\item{control.p}{NULL or named list of buckthorn control treatment parameters
set with \code{\link{set_control_p}}}

\item{verbose}{\code{TRUE} Show progress bar?}

\item{save_yrs}{\code{NULL} Vector of years to save; if \code{NULL}, all
years are returned}

\item{K_max}{\code{NULL} Maximum carrying capacity to avoid memory overloads}
}
\value{
Array N of abundances for each cell and age group, matrix B of seed
  bank abundances, matrix nFl with number of flowering individuals, matrix
  nSd of total seeds produced in each cell, matrix nSdStay of total seeds
  remaining in each cell, and matrix D of immigrants into each cell.
}
\description{
Run the simulation. Currently, it runs all time steps, but for the economic
model, the structure will need to be slightly adjusted to run a single time
step. The initialization is separated from this function for that reason.
}
\keyword{run,}
\keyword{simulate}
