% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_sim_lambda}
\alias{run_sim_lambda}
\title{Run lambda simulation}
\usage{
run_sim_lambda(ngrid, ncell, g.p, lambda, lc.df, sdd.pr, N.init,
  method = "wt.mn", verbose = F)
}
\arguments{
\item{ngrid}{Number of grid cells in entire map}

\item{ncell}{Number of inbound grid cells}

\item{g.p}{Named list of global parameters set with \code{\link{set_g_p}}}

\item{lambda}{Vector of length \code{n.lc} with lambdas for each land cover
type}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{sdd.pr}{Array with sdd probabilities and neighborhoods created by
\code{\link{sdd_set_probs}}}

\item{N.init}{Matrix or array with initial population sizes created by
\code{\link{pop_init}}}

\item{method}{\code{"wt.mn"} Method for calculating cell expectations, taking
values of \code{"wt.mn"} or \code{"lm"}. If \code{"wt.mn"}, the expectation
for each parameter is the weighted mean across land cover types
proportional to their coverage, with the land cover specific values stored
in the parameter vectors. If \code{"lm"}, the expectation is calculated in
a regression with the slopes contained in each parameter vector.
Individuals cannot be assigned to specific land cover categories with
\code{"lm"}, so \code{"m"} must be scalar.}

\item{verbose}{\code{TRUE} Show progress bar?}
}
\value{
Matrix N of abundances for each cell and time step and vector
  lambda.E of predicted lambda value for each cell
}
\description{
Run the simulation. Currently, it runs all time steps, but for the economic
model, the structure will need to be slightly adjusted to run a single time
step. The initialization is separated from this function for that reason.
}
\keyword{lambda}
\keyword{run,}
\keyword{simulate,}
