% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_dispersal.R
\name{sdd_set_probs}
\alias{sdd_set_probs}
\title{Set short distance dispersal probabilities}
\usage{
sdd_set_probs(ncell, lc.df, g.p, lc.new = NULL, edges = "wall",
  lc.col = 4:9, verbose = F)
}
\arguments{
\item{ncell}{Number of inbound grid cells}

\item{lc.df}{Dataframe or tibble with xy coords, land cover proportions, and
cell id info}

\item{g.p}{Named list of global parameters}

\item{lc.new}{Vector of indexes of cells that need to have their SDD
neighborhood recalculated; defaults to \code{NULL} and calculates SDD
neighborhoods for all cells}

\item{edges}{\code{"wall"} Boundary behavior}

\item{lc.col}{\code{4:9} Column indexes for land cover proportions}

\item{verbose}{\code{FALSE} Give updates for number of cells completed?}
}
\value{
List with full neighborhoods,i, and sparse representation, sp. The
  full array has dim(disp.rows, disp.cols, 2, ncell) where the third
  dimension contains grid id's for the neighborhood or probabilities to each
  target cell. The sparse representation is a list with an element for each
  cell, where each element is a vector of non-zero probabilities named with
  the target cells.
}
\description{
Assign base dispersal probabilities from each cell. Each layer \code{k} in
\code{[1:i, 1:j, k, source.id]} is the SDD neighborhood for cell n. k=1
contains pr(SDD | source.id,i,j); k=2 contains the grid id for each cell in
the neighborhood.
}
\keyword{dispersal,}
\keyword{probabilities}
\keyword{probability,}
\keyword{sdd,}
