% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{set_control_p}
\alias{set_control_p}
\title{Set buckthorn control treatment parameters}
\usage{
set_control_p(null_ctrl = TRUE, t.trt = 30, add.owners = FALSE,
  grd.i = NULL, man.i = NULL, chg.i = NULL, pTrt.grd = 0.05,
  pTrt.man = 0.05, grd.trt = c(Lit = 0.005, Cov = 0.01, Com = 1e-05),
  man.trt = c(M = 0.1, C = 0.3, MC = 0.8), lc.chg = TRUE, pChg = 1e-04)
}
\arguments{
\item{null_ctrl}{\code{TRUE} Set control parameters to \code{NULL}?}

\item{t.trt}{\code{30} Year to start treatments}

\item{add.owners}{\code{FALSE} Do owners treat every year once starting a
particular treatment?}

\item{grd.i}{\code{NULL} Vector of cell IDs to receive ground treatments. If
\code{NULL}, then \code{pTrt.grd * ncell} cells are assigned randomly with
the \link{trt_assign} function}

\item{man.i}{\code{NULL} Vector of cell IDs to receive manual treatments. If
\code{NULL}, then \code{pTrt.man * ncell} cells are assigned randomly with
the \link{trt_assign} function}

\item{chg.i}{\code{NULL} Vector of cell IDs to receive land cover changes. If
\code{NULL}, then \code{pChg * ncell} cells are assigned randomly with the
\link{trt_assign} function}

\item{pTrt.grd}{\code{0.05} Proportion of cells with ground treatments in
each time step; only used if \code{grd.i=NULL}}

\item{pTrt.man}{\code{0.05} Proportion of cells with manual treatments in
each time step; only used if \code{man.i=NULL}}

\item{grd.trt}{\code{c(Lit=0.005, Cov=0.01, Com=0.00001)} Named vector with
ground treatments and associated seedling establishment probabilities;
treatments include litter (Lit), cover crops (Cov), and soil compaction
(Com)}

\item{man.trt}{\code{c(M=0.1, C=0.3, MC=0.8)} Named vector with manual
treatments and associated mortality (=success) rates; treatments include
mechanical (M), chemical (C), or both (MC)}

\item{lc.chg}{\code{TRUE} Does land cover change (i.e., timber harvest)
across years?}

\item{pChg}{\code{0.0001} Proportion of cells with land cover change each
year}
}
\value{
Named list of control parameters including all arguments as elements
  unless \code{null_ctrl==TRUE}, in which case the function returns
  \code{NULL}
}
\description{
Set the buckthorn control treatment parameters for the simulation, allowing
for individual elements to be reassigned. Treatments fall into two broad
categories: ground cover and manual. Ground cover treatments reduce seedling
recruitment (i.e., establishment probabilities, \code{set_g_p()$p}) by
covering the ground with litter or cover crops, or by compacting the soil.
When these treatments are enacted in a cell, the establishment probability is
replaced by the corresponding probability assigned in
\code{set_control_p()$grd.trt}. Manual treatments reduce juvenile and adult
abundances by mechanical and/or chemical methods. The success rates (i.e.,
mortality rates) are assigned in \code{set_control_p()$man.trt}, and when a
manual treatment is enacted in a cell, the adult and juvenile abundances in
that cell are reduced by the corresponding rate.
}
\keyword{control,}
\keyword{initialize,}
\keyword{parameter}
\keyword{set}
\keyword{treatment,}
\keyword{up,}
