% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{set_control_p}
\alias{set_control_p}
\title{Set buckthorn control treatment parameters}
\usage{
set_control_p(null_ctrl = TRUE, t.trt = 30, add.owners = FALSE,
  grd.i = NULL, man.i = NULL, chg.i = NULL, pTrt.grd = 0.05,
  pTrt.man = 0.05, grd.trt = c(Lit = 0.005, Cov = 0.01, Com = 1e-05),
  man.trt = c(M = 0.1, C = 0.3, MC = 0.8), lc.chg = TRUE, pChg = 1e-04)
}
\arguments{
\item{null_ctrl}{\code{TRUE} Set control parameters to \code{NULL}?}

\item{t.trt}{\code{30} Year to start treatments}

\item{add.owners}{\code{FALSE} Do owners treat every year once starting a
particular treatment?}

\item{grd.i}{\code{NULL} Vector of cell IDs to receive ground treatments. If
\code{NULL}, then \code{pTrt.grd * ncell} cells are assigned randomly with
the \link{trt_assign} function}

\item{man.i}{\code{NULL} Vector of cell IDs to receive manual treatments. If
\code{NULL}, then \code{pTrt.man * ncell} cells are assigned randomly with
the \link{trt_assign} function}

\item{chg.i}{\code{NULL} Vector of cell IDs to receive land cover changes. If
\code{NULL}, then \code{pChg * ncell} cells are assigned randomly with
the \link{trt_assign} function}

\item{pTrt.grd}{\code{0.05} Proportion of cells with ground treatments in
each time step}

\item{pTrt.man}{\code{0.05} Proportion of cells with manual treatments in
each time step}

\item{grd.trt}{\code{Lit=0.005, Cov=0.01, Com=0.00001} Named vector with
ground treatments and associated seedling establishment probabilities}

\item{man.trt}{\code{c(M=0.1, C=0.3, MC=0.8)} Named vector with manual
treatments and associated mortality (=success) rates}

\item{lc.chg}{\code{TRUE} Does land cover change across years?}

\item{pChg}{\code{0.0001} Proportion of cells with land cover change each
year}
}
\value{
Named list of control parameters including all arguments as elements
  unless \code{null_ctrl==TRUE}, in which case the function returns
  \code{NULL}
}
\description{
This function sets the buckthorn control treatment parameters for the
simulation, allowing for individual elements to be reassigned.
}
\keyword{control,}
\keyword{initialize,}
\keyword{parameter}
\keyword{set}
\keyword{treatment,}
\keyword{up,}
