% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_aux.R
\name{set_g_p}
\alias{set_g_p}
\title{Set global parameters}
\usage{
set_g_p(tmax = 100, dem.st = FALSE, sdd.st = TRUE, bank = TRUE,
  n.cores = 4, lc.r = Inf, lc.c = Inf, n.lc = 6, N.p.t0 = 10,
  p.f = c(0.45, 0, 0.291, 0.309, 0.309, 0.272), mu = c(1948, 0, 14, 41,
  41, 21), gamma = 2.48, m = c(3, 3, 7, 7, 7, 7), p.c = c(0.165,
  0.165, 0.296, 0.252, 0.252, 0.296), sdd.rate = 0.03775, sdd.max = 24,
  bird.hab = c(0.32, 0.36, 0.05, 0.09, 0.09, 0.09), n.ldd = 19,
  s.c = 0.585, s.B = 0.72, s.M = c(0.9, 0, 0.6, 0.6, 0.6, 0.6),
  s.N = c(1, 1, 1, 1, 1, 1), K = c(28205, 0, 4162, 4162, 4162, 4162),
  g.D = 0, g.B = 0.2, p = c(0.295, 0, 0.421, 0.082, 0.082, 0.23),
  edges = "wall", method = "wt.mn")
}
\arguments{
\item{tmax}{\code{100} Number of time steps per simulation}

\item{dem.st}{\code{FALSE} Include stochasticity in demography?}

\item{sdd.st}{\code{TRUE} Include stochasticity in short distance dispersal?}

\item{bank}{\code{TRUE} Include seedbank?}

\item{n.cores}{\code{4} Number of cores for parallelizing sdd.pr calculation}

\item{lc.r}{\code{Inf} Maximum number of rows (\code{y}) in landscape}

\item{lc.c}{\code{Inf} Maximum number of columns (\code{x}) in landscape}

\item{n.lc}{\code{6} Number of land cover categories}

\item{N.p.t0}{\code{10} Number of cells with buckthorn at t=1}

\item{p.f}{\code{c(0.45, 0, 0.291, 0.309, 0.309, 0.272)} Vector
\code{length=n.lc} of fruiting probabilities}

\item{mu}{\code{c(1948, 0, 14, 41, 41, 21)} Vector \code{length=n.lc} of mean
fruit per adult}

\item{gamma}{\code{2.48} Scalar: mean number of seeds per fruit}

\item{m}{\code{c(3, 3, 7, 7, 7, 7)} Vector \code{length=n.lc} or scalar of
age at first fruiting. Individuals at this age are considered adults}

\item{p.c}{\code{c(0.165, 0.165, 0.296, 0.252, 0.252, 0.296)} Vector
\code{length=n.lc} of proportion of fruits eaten by birds, with
\code{1-p.c} assumed to drop directly below buckthorn individuals}

\item{sdd.rate}{\code{0.03775} 1/mn for exponential dispersal kernel (units =
cells); value assumes 20ac (8.1 ha) grid cells}

\item{sdd.max}{\code{24} Maximum dispersal distance in cells (units = cells);
value assumes 20ac (8.1 ha) grid cells}

\item{bird.hab}{\code{c(0.32, 0.36, 0.05, 0.09, 0.09, 0.09)} Vector
\code{length=n.lc} of relative bird habitat preferences}

\item{n.ldd}{\code{19} Number of long distance dispersal events per year}

\item{s.c}{\code{0.585} Seed viability post-digestion}

\item{s.B}{\code{0.72} Probability of annual survival in seed bank}

\item{s.M}{\code{c(0.9, 0. 0.6, 0.6, 0.6, 0.6)} Vector \code{length=n.lc} of
annual juvenile survival rates}

\item{s.N}{\code{c(1, 1, 1, 1, 1, 1)} Vector \code{length=n.lc} of annual
adult survival rates}

\item{K}{\code{c(28205, 0, 4162, 4162, 4162, 4162)} Vector (length=n.lc) of
carrying capacities for adults; values assume 20ac (8.1 ha) grid cells}

\item{g.D}{\code{0} Probability of direct germination (i.e., a seed
germinates in the same year it is produced)}

\item{g.B}{\code{0.2} Probability of germinating from the seed bank}

\item{p}{\code{c(0.295, 0, 0.421, 0.082, 0.082, 0.23)} Vector
\code{length=n.lc} of seedling establishment probabilities}

\item{edges}{\code{"wall"} Boundary behavior, taking values of \code{"wall"},
\code{"sink"}, or \code{"none"}. See boundary_behavior.Rmd for descriptions}

\item{method}{\code{"wt.mn"} Method for calculating cell expectations, taking
values of \code{"wt.mn"} or \code{"lm"}. If \code{"wt.mn"}, the expectation
for each parameter is the weighted mean across land cover types
proportional to their coverage, with the land cover specific values stored
in the parameter vectors. If \code{"lm"}, the expectation is calculated in
a regression with the slopes contained in each parameter vector.
Individuals cannot be assigned to specific land cover categories with
\code{"lm"}, so \code{"m"} must be scalar.}
}
\value{
Named list of global parameters including all arguments as elements
}
\description{
Set the global parameters for the simulation, allowing for individual
elements to be reassigned. Default parameters assume 8.1 hectare cells.
}
\keyword{global,}
\keyword{initialize,}
\keyword{parameter}
\keyword{set}
\keyword{up,}
